+++ cocos/physics3d/CCPhysics3DComponent.cpp
@@ -100,7 +100,7 @@ void Physics3DComponent::addToPhysicsWorld(Physics3DWorld* world)
         {
             auto parent = _owner->getParent();
             while (parent) {
-                for (int i = 0; i < components.size(); i++) {
+                for (size_t i = 0; i < components.size(); i++) {
                     if (parent == components[i]->getOwner())
                     {
                         //insert it here
