+++ tests/cpp-tests/Classes/ExtensionsTest/TableViewTest/TableViewTestScene.cpp
@@ -47,7 +47,7 @@ bool TableViewTest::init()
 
 void TableViewTest::tableCellTouched(TableView* table, TableViewCell* cell)
 {
-    CCLOG("cell touched at index: %ld", cell->getIdx());
+    CCLOG("cell touched at index: %ld", static_cast<long>(cell->getIdx()));
 }
 
 Size TableViewTest::tableCellSizeForIndex(TableView *table, ssize_t idx)
@@ -60,7 +60,7 @@ Size TableViewTest::tableCellSizeForIndex(TableView *table, ssize_t idx)
 
 TableViewCell* TableViewTest::tableCellAtIndex(TableView *table, ssize_t idx)
 {
-    auto string = StringUtils::format("%ld", idx);
+    auto string = StringUtils::format("%ld", static_cast<long>(idx));
     TableViewCell *cell = table->dequeueCell();
     if (!cell) {
         cell = new (std::nothrow) CustomTableViewCell();
